requirements.txt
maxapi==1.0.0
asyncpg
pytz
python-dotenv